-extracted Teams types -replaced header prop in the ColorBox with copyToClipboardIcon
