+++ packages/react/src/themes/types.ts
@@ -165,38 +165,6 @@ export type ColorSchemeMappingOverrides<
   default?: Partial<Extendable<ColorScheme, string>>
 }
 
-/**
- * Teams theme specific interfaces
- */
-export type TeamsContextualColors = {
-  brand: ColorVariants
-}
-
-export type TeamsNaturalColors = {
-  grey: ColorVariants
-  green: ColorVariants
-  orange: ColorVariants
-  red: ColorVariants
-  yellow: ColorVariants
-  pink: ColorVariants
-}
-
-export type TeamsTransparentColors = {
-  silver: ColorVariants
-  ruby: ColorVariants
-  onyx: ColorVariants
-}
-
-export type TeamsColorNames = keyof (TeamsContextualColors &
-  TeamsNaturalColors &
-  PrimitiveColors &
-  TeamsTransparentColors)
-
-export type TeamsSchemeMappingWithAreas<TAreas extends string> = StrictColorSchemeMapping<
-  StrictColorScheme<TAreas>,
-  TeamsColorNames
->
-
 export type ItemType<T> = T extends (infer TItem)[] ? TItem : never
 
 // ========================================================
